Update generators, create new schemas
