+++ docs/hub/.schema/linterhub.version.md
@@ -1,0 +1,14 @@
+# Linterhub Version
+This document describes the structure of linterhub version output
+## Structure
+Version of linterhub
+
+|Key|Type|Required|Description|
+|-|:-:|:-:|-|
+|version|string|+|Version of linterhub|
+## Example
+```
+{
+    "version": "string"
+}
+```
