+++ src/hub/.schema/engine.version.json
@@ -1,0 +1,44 @@
+{
+    "$schema": "http://json-schema.org/draft-04/schema",
+    "title": "This document describes the structure of engine version output",
+    "description": "Engine version output is an array of objects describing the state of the engines",
+    "type": "array",
+    "items": {
+        "$ref": "#/definitions/result"
+    },
+    "definitions": {
+        "result": {
+            "type": "object",
+            "description": "Represents install/version request result",
+            "properties": {
+                "installed": {
+                    "type": "boolean",
+                    "description": "Installation status"
+                },
+                "message": {
+                    "type": "string",
+                    "description": "Error message if cant install engine"
+                },
+                "name": {
+                    "type": "string",
+                    "description": "Engine name"
+                },
+                "version": {
+                    "type": "string",
+                    "description": "Engine version"
+                },
+                "packages": {
+                    "type": "array",
+                    "description": "List of dependencies",
+                    "items": {
+                        "$ref": "#/definitions/result"
+                    }
+                }
+            },
+            "required": [
+                "name",
+                "installed"
+            ]
+        }
+    }
+}
