--- src/hub/.schema/linterhub.output.json
@@ -1,54 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-04/schema",
-    "title": "This document describes the structure of linterhub output",
-    "description": "Linterhub output is an array of engines results",
-    "type": "array",
-    "items": {
-        "$ref": "#/definitions/result"
-    },
-    "definitions": {
-        "result": {
-            "type": "object",
-            "description": "The engine result",
-            "properties": {
-                "engine": {
-                    "type": "string",
-                    "description": "The engine name that performed analysis"
-                },
-                "results": {
-                    "type": "array",
-                    "description": "The list of analysis results",
-                    "items": {
-                        "$ref": "engine.output.json"
-                    }
-                },
-                "error": {
-                    "type": "object",
-                    "description": "The problem definition if analysis is not possible",
-                    "properties": {
-                        "code": {
-                            "type": "integer",
-                            "description": "The error code"
-                        },
-                        "title": {
-                            "type": "string",
-                            "description": "The error title"
-                        },
-                        "description": {
-                            "type": "string",
-                            "description": "The error decription"
-                        }
-                    },
-                    "required": [
-                        "code",
-                        "title"
-                    ]
-                }
-            },
-            "required": [
-                "engine",
-                "results"
-            ]
-        }
-    }
-}
